authentication changes

+++ src/server/db/resolvers.js
@@ -1,18 +1,72 @@
-const Sequelize = require("sequelize");
+//const Sequelize = require("sequelize");
+import Sequelize from "sequelize";
+import PasswordHash from "password-hash";
 
-const resolvers = {
+const isAuthenticated = async (models, headers) => {
+    let session = await models.Session.findOne({
+        where: { userAgent: headers["user-agent"] },
+    });
+    if (session) {
+        return Boolean(session.authenticated);
+    }
+    return false;
+};
+
+export const resolvers = {
     Query: {
-        users(_, { input }, { models }) {
+        async login(_, { username, password }, { models, headers }) {
+            let user = await models.User.findOne({ where: { username } });
+            if (user) {
+                if (PasswordHash.verify(password, user.password)) {
+                    models.Session.create({
+                        userAgent: headers["user-agent"],
+                        authenticated: true,
+                        authenticatedUser: user.id,
+                    });
+                    return true;
+                }
+            }
+            return false;
+        },
+        async logout(_, {}, { models, headers }) {
+            if (!(await isAuthenticated(models, headers))) return null;
+            return models.Session.destroy({
+                where: { userAgent: headers["user-agent"] },
+            });
+        },
+        async authenticated(_, {}, { models, headers }) {
+            let session = await models.Session.findOne({
+                where: { userAgent: headers["user-agent"] },
+            });
+            if (session) {
+                return Boolean(session.authenticated);
+            }
+            return false;
+        },
+        async authenticatedUser(_, {}, { models, headers }) {
+            if (!(await isAuthenticated(models, headers))) return null;
+            let session = await models.Session.findOne({
+                where: { userAgent: headers["user-agent"] },
+            });
+            if (session) {
+                return models.User.findByPk(session.authenticatedUser);
+            }
+        },
+        async users(_, { input }, { models, headers }) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.User.findAll();
         },
-        user(_, { id }, { models }) {
+        async user(_, { id }, { models, headers }) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.User.findByPk(id);
         },
-        userByUsername(_, { username }, { models }) {
+        async userByUsername(_, { username }, { models, headers }) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.User.findOne({ where: { username } });
         },
 
-        groceries(_, { name }, { models }) {
+        async groceries(_, { name }, { models, headers }) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.Grocery.findAll({
                 where: {
                     ...(name
@@ -29,26 +83,39 @@ const resolvers = {
                 ],
             });
         },
-        grocery(_, { id }, { models }) {
+        async grocery(_, { id }, { models, headers }) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.Grocery.findByPk(id);
         },
 
-        calendarItems(_, { input }, { models }) {
+        async calendarItems(_, { input }, { models, headers }) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.CalendarItem.findAll();
         },
-        calendarItem(_, { id }, { models }) {
+        async calendarItem(_, { id }, { models, headers }) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.CalendarItem.findByPk(id);
         },
     },
     Mutation: {
-        async createUser(root, { id, username, password, color }, { models }) {
+        async createUser(
+            root,
+            { id, username, password, color },
+            { models, headers }
+        ) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.User.create({
                 username,
                 password,
                 color,
             });
         },
-        async updateUser(root, { id, username, password, color }, { models }) {
+        async updateUser(
+            root,
+            { id, username, password, color },
+            { models, headers }
+        ) {
+            if (!(await isAuthenticated(models, headers))) return null;
             await models.User.update(
                 { username, password, color },
                 { where: { id } }
@@ -59,15 +126,21 @@ const resolvers = {
         async updateGrocery(
             root,
             { id, name, amount, checked, user, updated_by_user },
-            { models }
+            { models, headers }
         ) {
+            if (!(await isAuthenticated(models, headers))) return null;
             await models.Grocery.update(
                 { name, amount, checked, user, updated_by_user },
                 { where: { id } }
             );
             return models.Grocery.findByPk(id);
         },
-        async createGrocery(root, { name, amount, checked, user }, { models }) {
+        async createGrocery(
+            root,
+            { name, amount, checked, user },
+            { models, headers }
+        ) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.Grocery.create({
                 name,
                 amount,
@@ -75,7 +148,8 @@ const resolvers = {
                 user,
             });
         },
-        async deleteCheckedGroceries(root, {}, { models }) {
+        async deleteCheckedGroceries(root, {}, { models, headers }) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.Grocery.destroy({
                 where: {
                     checked: true,
@@ -83,7 +157,8 @@ const resolvers = {
             });
         },
 
-        async deleteGrocery(root, { id }, { models }) {
+        async deleteGrocery(root, { id }, { models, headers }) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.Grocery.destroy({
                 where: {
                     id,
@@ -94,8 +169,9 @@ const resolvers = {
         async createCalendarItem(
             root,
             { title, description, startDateTime, endDateTime, user },
-            { models }
+            { models, headers }
         ) {
+            if (!(await isAuthenticated(models, headers))) return null;
             return models.CalendarItem.create({
                 title,
                 description,
@@ -107,14 +183,17 @@ const resolvers = {
     },
     CalendarItem: {
         async user(user, args, ctx, info) {
+            //if (!(await isAuthenticated(models, headers))) return null;
             return resolvers.Query.user(null, { id: user.user }, ctx);
         },
     },
     Grocery: {
         async user(user, args, ctx, info) {
+            //if (!(await isAuthenticated(models, headers))) return null;
             return resolvers.Query.user(null, { id: user.user }, ctx);
         },
         async updated_by_user(updated_by_user, args, ctx, info) {
+            //if (!(await isAuthenticated(models, headers))) return null;
             return resolvers.Query.user(
                 null,
                 { id: updated_by_user.updated_by_user },
@@ -124,4 +203,4 @@ const resolvers = {
     },
 };
 
-module.exports = resolvers;
+//module.exports = resolvers;
